decision: refactoring tooltips
The informations feeding the tooltip of an activity were collected and
combined by the DrawingArea. The tooltips system was not used the
wright way. From now on, the tooltips is used the proper way, and
methods have been added in order to delegate information query to the
dedicated classes representing the Model part of the MVC design
pattern.

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>