decision: add a relative date to year concept
The plan file accepts relative calendar date like this "+YYYY-MM-DD"
where +YYYY stands for the number of the year. For exampple "+0-3-10"
stands for the third of february of the year n of the load date of the
plan. And "+1-11-5" stands of the fifth november of the year n+1. The
29 february is not allowed. Reklative numerical date are still
allowed.  Inside the modeling plugin. The Activity Dialog window allow
this new feature.