gvle2.simulating.shape: add a plugin
This plugin enable to draw a shapeFile on a window what's become
available inside storage views. The plugin expect to find a port
parameter named "GisFileName" that is valued with the name of a
shapeFile. The plugin does use a single storage view, expecting to
find values of a single state variable in each column. Each column
should contain the "color" value of a shape during the simulation. A
color table is settable inside the gvle2.conf file.