gvle.discrete-time: add a new plugin
The last but not the least proposition, and not the final.  More a
POC, the user workflow will be simplified.  We have decided to edit
the files at the top level, and directly.  In the next version we are
going to use the central version of the main window, but also tge
right par.  This plugin still validate undo/redo save, and meta-data
usage.