decision: enable to relax controls on resources
From now on, if a resource constraint needs more resources than
defined, a warning is sent to the vle.log file. But, a new port
parameter is available is provided to raise an exception insteed of
sending a warning. This is a boolean parameter called kbResourcesChek,
and when set to true, the exception is raised.